ARChatBot/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# RAG / Embeddings
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-core>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0

# LLM & routing
groq>=0.9.0

# Database
pymongo>=4.7.0

# Data analytics
pandas>=2.2.0

# Config / env
python-dotenv>=1.0.0
